Reduce public interface of GenericEvent
Mark many methods which should not be used by the
analysis user as protected. The public interface of GenericEvent
is now essentially restricted to get/set and is_valid.
As many classes in the framework need access to those, introduce the EventAccess_
class, which can be used for that. Note that this class should
never be used in analysis code.