Add L1Prefire weights module to ntuple for 2016/7
From
https://twiki.cern.ch/twiki/bin/viewauth/CMS/L1ECALPrefiringWeightRecipe

This stores prefire weight, and up & down variations for 2016/7 MC.
The handling of the ROOT file location is not optimal, should
be changed centrally at some point.

Note that weight branches are made even for data & 2018 MC,
to make life easier in EventHelper. But maybe this is a waste
of space, and we need to be more clever.