Save year arg in generate_process() to Event class
Analysis modules can then dynamically change their behaviour
depending on this parameter, or the user can choose how to utilise it.

string was chosen instead of an int, because the latter relies on
some immutable mapping (or rather, never changing the order of
existing values). This way, the user can immediately inspect
and has more control over how to react.
1 string is also a very small overhead.