MET::mEtSig -> mEtSignificance, use proper significance
Previously had stored mEtSig, which was MET/sqrt(sumEt).
Since this is still available by using sumEt(), we now
also store the significance as calcualted from the covariance matrix.

Note that this is not possible for the Gen Met, since the significance
matrix is empty. We'd have to recalculate that explicitly, which
may be more trouble than it's worth. So for now, that keeps
storing MET/sqrt(sumEt).