Redo DeepBtag/flav + PUPPI multi to simplify, apply to reclustered AK8PUPPI
Now has similar functionality as in #1099, but now using our reclustered
AK8PUPPI fat jets for substructure instead of slimmedJetsAK8.

Rejiggle modules, and try to setup a bit cleaner to avoid multiple calls
to updateJetCollection. The trick is to do PUPPI multiplicities last,
since we use the collection name to lookup the userfloat for it.

Also include sneaky postfix required for deep btag for reclustered PUPPI jets.